% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.alg.R
\name{em.alg}
\alias{em.alg}
\title{Internal function for performing EM algorithm}
\usage{
em.alg(longdat, survdat, model, ran, lat, sepassoc, paraests, gpt, max.it, tol)
}
\description{
Internal function for performing EM algorithm
}
\keyword{internal}

