% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlike.R
\name{jlike}
\alias{jlike}
\title{Internal function for calculating the likelihood after the maximizer has been
determined from the EM algorithm}
\usage{
jlike(longdat, survdat, model, ran, lat, sepassoc, likeests, lgpt)
}
\description{
Internal function for calculating the likelihood after the maximizer has been
determined from the EM algorithm
}
\keyword{internal}

