% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joineR-package.R
\docType{package}
\name{joineR-package}
\alias{joineR-package}
\title{joineR: Joint Modelling of Repeated Measurements and Time-to-Event Data}
\description{
Analysis of repeated measurements and time-to-event data via random effects joint models. Fits the joint models proposed by Henderson and colleagues \doi{10.1093/biostatistics/1.4.465} (single event time) and by Williamson and colleagues (2008) \doi{10.1002/sim.3451} (competing risks events time) to a single continuous repeated measure. The time-to-event data is modelled using a (cause-specific) Cox proportional hazards regression model with time-varying covariates. The longitudinal outcome is modelled using a linear mixed effects model. The association is captured by a latent Gaussian process. The model is estimated using am Expectation Maximization algorithm. Some plotting functions and the variogram are also included. This project is funded by the Medical Research Council (Grant numbers G0400615 and MR/M013227/1).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/graemeleehickey/joineR/}
  \item Report bugs at \url{https://github.com/graemeleehickey/joineR/issues}
}

}
\author{
\strong{Maintainer}: Graeme L. Hickey \email{graemeleehickey@gmail.com} (\href{https://orcid.org/0000-0002-4989-0054}{ORCID})

Authors:
\itemize{
  \item Pete Philipson \email{peter.philipson1@newcastle.ac.uk} (\href{https://orcid.org/0000-0001-7846-0208}{ORCID})
  \item Ines Sousa \email{isousa@mct.uminho.pt} (\href{https://orcid.org/0000-0002-2712-1713}{ORCID})
  \item Peter J. Diggle \email{p.diggle@lancaster.ac.uk} (\href{https://orcid.org/0000-0003-3521-5020}{ORCID})
  \item Paula Williamson \email{p.r.williamson@liverpool.ac.uk} (\href{https://orcid.org/0000-0001-9802-6636}{ORCID})
  \item Ruwanthi Kolamunnage-Dona \email{kdrr@liverpool.ac.uk} (\href{https://orcid.org/0000-0003-3886-6208}{ORCID})
  \item Robin Henderson \email{robin.henderson@ncl.ac.uk}
}

Other contributors:
\itemize{
  \item Maria Sudell \email{m.e.sudell@liverpool.ac.uk} [contributor]
  \item Medical Research Council (Grant numbers: G0400615 and MR/M013227/1) [funder]
}

}
\keyword{internal}
