% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longst.R
\name{longst}
\alias{longst}
\title{Internal function to get initial parameter estimates for the longitudinal
sub-model}
\usage{
longst(longdat, long.formula, model, longdat2)
}
\description{
Internal function to get initial parameter estimates for the longitudinal
sub-model
}
\keyword{internal}

