% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sep}
\alias{sep}
\title{Internal function to evaluate whether separate model fits should be reported}
\usage{
sep(ests, logical)
}
\description{
Internal function to evaluate whether separate model fits should be reported
}
\keyword{internal}

