% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simjoint.R
\name{simjoint}
\alias{simjoint}
\title{Simulate data from a joint model}
\usage{
simjoint(n = 500, model = c("intslope", "int", "quad"), sepassoc = FALSE,
  ntms = 5, b1 = c(1, 1, 1, 1), b2 = c(1, 1), gamma = c(1, 0.1), sigu,
  vare = 0.01, theta0 = -3, theta1 = 1, censoring = TRUE,
  censlam = exp(-3), truncation = FALSE, trunctime = max(ntms),
  gridstep = 0.01)
}
\arguments{
\item{n}{the number of subjects to simulate data for.}

\item{model}{a character string specifying the type of latent association. 
This defaults to the intercept and slope version as seen in Wulfsohn and 
Tsiatis (1997). For association via the random intercept only, choose
\code{model = "int"}, whereas for a quadratic association, use \code{model
= "quad"}. Computing times are commensurate with the type of association
structure chosen.}

\item{sepassoc}{logical value: if \code{TRUE} then the joint model is fitted 
with separate association, see \code{Details}.}

\item{ntms}{the maximum number of (discrete) time points to simulate repeated
longitudinal measurements at.}

\item{b1}{a vector specifying the coefficients of the fixed effects in the
longitudinal sub-model. The order in each row is intercept, a continuous 
covariate, covariate, a binary covariate, the measurement time.}

\item{b2}{a vector of \code{length = 2} specifying the coefficients for the
time-to-event baseline covariates, in the order of a continuous covariate
and a binary covariate.}

\item{gamma}{a vector of specifying the latent association parameter(s) for
the longitudinal outcome. It must be of length 1 if \code{sepassoc =
FALSE}.}

\item{sigu}{a positive-definite matrix specifying the variance-covariance 
matrix. If \code{model = "int"}, the matrix has dimension \code{dim = c(1,
1)}; if \code{model = "intslope"}, the matrix has dimension \code{dim = 
c(2, 2)}; else if \code{model = "quad"}, the matrix has dimension \code{dim
= c(3, 3)}. If \code{D = NULL} (default), an identity matrix is assumed.}

\item{vare}{a numeric value specifying the residual standard error.}

\item{theta0}{the log-scale parameter for a Gompertz distribution used to 
simulate the time-to-event outcome. The scale is calculated as 
\eqn{\exp(\theta_0)} to ensure it is positive.}

\item{theta1}{the shape parameter for a Gompertz distribution used to 
simulate the time-to-event outcome.}

\item{censoring}{logical: if \code{TRUE}, includes an independent censoring 
time.}

\item{censlam}{a scale (\eqn{>0}) parameter for an exponential distribution 
used to simulate random censoring times for when \code{censoring = TRUE}.}

\item{truncation}{logical: if \code{TRUE}, adds a truncation time for a 
maximum event time.}

\item{trunctime}{a truncation time for use when \code{truncation = TRUE}.}

\item{gridstep}{the step-size for the grid used to simulate event times when
\code{model = "quad"}. Default is \code{gridstep = 0.01}.}
}
\value{
A list of 2 data.frames: one recording the requisite longitudinal 
  outcomes data, and one recording the time-to-event data.
}
\description{
This function simulates longitudinal and time-to-event data from a joint
model.
}
\details{
The function \code{simData} simulates data from a joint model, 
  similar to that performed in Henderson et al. (2000). It works by first 
  simulating longitudinal data for all possible follow-up times using random
  draws for the multivariate Gaussian random effects and residual error
  terms. Data can be simulated assuming either random-intercepts only in each
  of the longitudinal sub-models; random-intercepts and random-slopes; or
  quadratic random effects structures. The failure times are simulated from
  proportional hazards time-to-event models; either an exponential
  distribution (in the case of random-intercepts models) or Gompertz
  distribution (random-intercept and random-slopes models) conditional on
  either known baseline effects (\code{model = "int"}), or a predictable
  time-varying process (\code{model = "intslope"}). In the case of the
  former, the methodology of Bender et al. (2005) is used to simulate the
  time, whilst in the case of latter, the approach of Austin (2012) is used.
  
  TODO: add something on the \code{model = "quad"}. Also, check
  \code{grid.step} defintion.
}
\examples{
simjoint(10, sepassoc = TRUE)
}
\author{
Pete Philipson (\email{pete.philipson@northumbria.ac.uk})
}
\references{
Austin PC. Generating survival times to simulate Cox proportional hazards 
models with time-varying covariates. \emph{Stat Med.} 2012; \strong{31(29)}: 
3946-3958.

Bender R, Augustin T, Blettner M. Generating survival times to simulate Cox 
proportional hazards models. \emph{Stat Med.} 2005; \strong{24}: 1713-1723.

Henderson R, Diggle PJ, Dobson A. Joint modelling of longitudinal 
measurements and event time data. \emph{Biostatistics.} 2000; \strong{1(4)}: 
465-480.
}
\keyword{datagen}
\keyword{survival}

