% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cjoint.R
\name{summary.cjoint}
\alias{summary.cjoint}
\title{Summarise a random effects joint model fit}
\usage{
\method{summary}{cjoint}(object, variance = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{cjoint}.}

\item{variance}{should the variance components be output as variances or
standard deviations? Defaults to \code{variance = TRUE}.}

\item{...}{further arguments for the summary.}
}
\value{
An object inheriting from class \code{summary.cjoint} with all
  components included in \code{object} (see \code{\link{cjoint}} for a full
  description of the components) plus the following components:

  \item{\code{nobs}}{the total number of (typically longitudinal)
  observations (i.e. rows in an unbalanced data set).}

  \item{\code{ngrps}}{the number of groups in the analyzed dataset, often
  individual subjects.}
}
\description{
Generic function used to produce summary information from a fitted random
effects joint model as represented by \code{object} of class \code{cjoint}.
}
\author{
Pete Philipson (\email{pete.philipson@northmbria.ac.uk}) and Graeme
  L. Hickey (\email{graeme.hickey@liverpool.ac.uk})
}
\keyword{methods}
