% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survst.R
\name{survst}
\alias{survst}
\title{Internal function for calculating the initial parameters of the survival data
sub-model}
\usage{
survst(survdat, surv.formula, survdat2)
}
\description{
Internal function for calculating the initial parameters of the survival data
sub-model
}
\keyword{internal}

