% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlike.R
\name{jlike}
\alias{jlike}
\title{Internal function for calculating the log-likelihood the maximizer}
\usage{
jlike(longdat, survdat, model, ran, lat, sepassoc, likeests, lgpt)
}
\description{
Internal function for calculating the log-likelihood the maximizer
}
\keyword{internal}

