% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longst.R
\name{longst}
\alias{longst}
\title{Internal function for calculating the initial parameters of the longitudinal
data sub-model}
\usage{
longst(longdat, long.formula, model, longdat2)
}
\description{
Internal function for calculating the initial parameters of the longitudinal
data sub-model
}
\keyword{internal}

