% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survst.R
\name{survst}
\alias{survst}
\title{Internal function to get initial parameter estimates for the time-to-event 
sub-model}
\usage{
survst(survdat, surv.formula, survdat2)
}
\description{
Internal function to get initial parameter estimates for the time-to-event 
sub-model
}
\keyword{internal}

